Rename detail.rs testcase into general.rs
